fix link detach vendor error exception parity
