+++ sdk/eventhub/azure-eventhub/tests/livetest/asynctests/test_negative_async.py
@@ -12,13 +12,14 @@ from azure.eventhub import (
     EventData,
     EventDataBatch,
 )
+from azure.identity.aio import EnvironmentCredential
 from azure.eventhub.exceptions import (
     EventHubError,
     ConnectError,
     AuthenticationError,
-    EventDataSendError
 )
-from azure.eventhub.aio import EventHubConsumerClient, EventHubProducerClient
+from azure.eventhub.aio import EventHubConsumerClient, EventHubProducerClient, EventHubSharedKeyCredential
+from azure.eventhub.aio._client_base_async import EventHubSASTokenCredential
 
 
 @pytest.mark.liveTest
@@ -198,3 +199,158 @@ async def test_invalid_proxy_server(connection_str, uamqp_transport):
     async with client:
         with pytest.raises(EventHubError):
             batch = await client.create_batch()
+
+@pytest.mark.liveTest
+@pytest.mark.asyncio
+async def test_client_invalid_credential_async(live_eventhub, uamqp_transport):
+
+    async def on_event(partition_context, event):
+        pass
+
+    async def on_error(partition_context, error):
+        on_error.err = error
+
+    env_credential = EnvironmentCredential()
+    producer_client = EventHubProducerClient(fully_qualified_namespace="fakeeventhub.servicebus.windows.net",
+                                             eventhub_name=live_eventhub['event_hub'],
+                                             credential=env_credential,
+                                             user_agent='customized information',
+                                             retry_total=1,
+                                             retry_mode='exponential',
+                                             retry_backoff=0.02,
+                                             uamqp_transport=uamqp_transport)
+    consumer_client = EventHubConsumerClient(fully_qualified_namespace="fakeeventhub.servicebus.windows.net",
+                                             eventhub_name=live_eventhub['event_hub'],
+                                             credential=env_credential,
+                                             user_agent='customized information',
+                                             consumer_group='$Default',
+                                             retry_total=1,
+                                             retry_mode='exponential',
+                                             retry_backoff=0.02,
+                                             uamqp_transport=uamqp_transport)
+    async with producer_client:
+        with pytest.raises(ConnectError):
+            await producer_client.create_batch(partition_id='0')
+
+    on_error.err = None
+    async with consumer_client:
+        task = asyncio.ensure_future(consumer_client.receive(on_event,
+                                                                starting_position= "-1", on_error=on_error))
+        await asyncio.sleep(15)
+    await task
+    assert isinstance(on_error.err, ConnectError)
+
+    producer_client = EventHubProducerClient(fully_qualified_namespace=live_eventhub['hostname'],
+                                             eventhub_name='fakehub',
+                                             credential=env_credential,
+                                             uamqp_transport=uamqp_transport)
+
+    consumer_client = EventHubConsumerClient(fully_qualified_namespace=live_eventhub['hostname'],
+                                             eventhub_name='fakehub',
+                                             credential=env_credential,
+                                             consumer_group='$Default',
+                                             retry_total=0,
+                                             uamqp_transport=uamqp_transport)
+
+    async with producer_client:
+        with pytest.raises(ConnectError):
+            await producer_client.create_batch(partition_id='0')
+
+    on_error.err = None
+    async with consumer_client:
+        task = asyncio.ensure_future(consumer_client.receive(on_event,
+                                                                starting_position= "-1", on_error=on_error))
+        await asyncio.sleep(15)
+    await task
+    assert isinstance(on_error.err, AuthenticationError)
+
+    credential = EventHubSharedKeyCredential(live_eventhub['key_name'], live_eventhub['access_key'])
+    auth_uri = "sb://{}/{}".format(live_eventhub['hostname'], live_eventhub['event_hub'])
+    token = (await credential.get_token(auth_uri)).token
+    producer_client = EventHubProducerClient(fully_qualified_namespace=live_eventhub['hostname'],
+                                             eventhub_name=live_eventhub['event_hub'],
+                                             credential=EventHubSASTokenCredential(token[:-1], time.time() + 5),
+                                             uamqp_transport=uamqp_transport)
+    await asyncio.sleep(6)
+    # expired credential
+    async with producer_client:
+        with pytest.raises(AuthenticationError):
+            await producer_client.create_batch(partition_id='0')
+
+    consumer_client = EventHubConsumerClient(fully_qualified_namespace=live_eventhub['hostname'],
+                                             eventhub_name=live_eventhub['event_hub'],
+                                             credential=EventHubSASTokenCredential(token, time.time() + 7),
+                                             consumer_group='$Default',
+                                             retry_total=0,
+                                             uamqp_transport=uamqp_transport)
+    on_error.err = None
+    async with consumer_client:
+        task = asyncio.ensure_future(consumer_client.receive(on_event,
+                                                                starting_position= "-1", on_error=on_error))
+        await asyncio.sleep(15)
+    await task
+
+    # expired credential
+    assert isinstance(on_error.err, AuthenticationError)
+
+    credential = EventHubSharedKeyCredential('fakekey', live_eventhub['access_key'])
+    producer_client = EventHubProducerClient(fully_qualified_namespace=live_eventhub['hostname'],
+                                             eventhub_name=live_eventhub['event_hub'],
+                                             credential=credential,
+                                             uamqp_transport=uamqp_transport)
+
+    async with producer_client:
+        with pytest.raises(AuthenticationError):
+            await producer_client.create_batch(partition_id='0')
+
+    producer_client = EventHubProducerClient(fully_qualified_namespace=live_eventhub['hostname'],
+                                             eventhub_name=live_eventhub['event_hub'],
+                                             credential=env_credential,
+                                             connection_verify="cacert.pem",
+                                             uamqp_transport=uamqp_transport)
+
+    # TODO: this seems like a bug from uamqp, should be ConnectError?
+    async with producer_client:
+        with pytest.raises(EventHubError):
+            await producer_client.create_batch(partition_id='0')
+
+    consumer_client = EventHubConsumerClient(fully_qualified_namespace=live_eventhub['hostname'],
+                                             eventhub_name=live_eventhub['event_hub'],
+                                             consumer_group='$Default',
+                                             credential=env_credential,
+                                             retry_total=0,
+                                             connection_verify="cacert.pem",
+                                             uamqp_transport=uamqp_transport)
+    async with consumer_client:
+        task = asyncio.ensure_future(consumer_client.receive(on_event,
+                                                                starting_position= "-1", on_error=on_error))
+        await asyncio.sleep(5)
+    await task
+
+    # TODO: this seems like a bug from uamqp, should be ConnectError?
+    assert isinstance(on_error.err, FileNotFoundError)
+
+    producer_client = EventHubProducerClient(fully_qualified_namespace=live_eventhub['hostname'],
+                                             eventhub_name=live_eventhub['event_hub'],
+                                             credential=env_credential,
+                                             custom_endpoint_address="fakeaddr",
+                                             uamqp_transport=uamqp_transport)
+
+    async with producer_client:
+        with pytest.raises(AuthenticationError):
+            await producer_client.create_batch(partition_id='0')
+
+    consumer_client = EventHubConsumerClient(fully_qualified_namespace=live_eventhub['hostname'],
+                                             eventhub_name=live_eventhub['event_hub'],
+                                             consumer_group='$Default',
+                                             credential=env_credential,
+                                             retry_total=0,
+                                             custom_endpoint_address="fakeaddr",
+                                             uamqp_transport=uamqp_transport)
+    async with consumer_client:
+        task = asyncio.ensure_future(consumer_client.receive(on_event,
+                                                                starting_position= "-1", on_error=on_error))
+        await asyncio.sleep(15)
+    await task
+
+    assert isinstance(on_error.err, AuthenticationError)
